requirements.txt
asgiref==3.5.1
astroid==2.11.5
autopep8==1.6.0
dill==0.3.4
Django==4.0.4
django-dotenv==1.4.2
djangorestframework==3.13.1
flake8==4.0.1
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.6.1
platformdirs==2.5.2
psycopg2==2.9.3
pycodestyle==2.8.0
pyflakes==2.4.0
pytz==2022.1
sqlparse==0.4.2
toml==0.10.2